move performance tests to separate file
